Crate i18n_config

This library contains the configuration structs (along with their parsing functions) for the cargo-i18n tool/system.

Structs

Crate

Represents a rust crate.

CratePaths

Important i18n-config paths related to the current crate.

FluentConfig

The data structure representing what is stored (and possible to store) within the fluent subsection of a i18n.toml file.

GettextConfig

The data structure representing what is stored (and possible to store) within the gettext subsection of a i18n.toml file.

I18nConfig

The data structure representing what is stored (and possible to store) within a i18n.toml file.

Enums

I18nConfigError

An error type for use with the i18n-config crate.

WhyNotCrate

An error type explaining why a crate failed to verify.

Functions

locate_crate_paths

Locate the current crate’s directory and i18n.toml config file. This is intended to be called by a procedural macro during crate compilation.